org.gamegineer.table.net/src/org/gamegineer/table/net/NetworkTableFactory.java
/*
 * NetworkTableFactory.java
 * Copyright 2008-2010 Gamegineer.org
 * All rights reserved.
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 *
 * Created on Nov 4, 2010 at 11:44:35 PM.
 */

package org.gamegineer.table.net;

import net.jcip.annotations.ThreadSafe;
import org.gamegineer.table.core.ITable;
import org.gamegineer.table.internal.net.NetworkTable;

/**
 * A factory for creating network tables.
 */
@ThreadSafe
public final class NetworkTableFactory
{
    // ======================================================================
    // Constructors
    // ======================================================================

    /**
     * Initializes a new instance of the {@code NetworkTableFactory} class.
     */
    private NetworkTableFactory()
    {
        super();
    }


    // ======================================================================
    // Methods
    // ======================================================================

    /**
     * Creates a new network table.
     * 
     * @param table
     *        The table to be attached to the network; must not be {@code null}.
     * 
     * @return A new network table; never {@code null}.
     * 
     * @throws java.lang.NullPointerException
     *         If {@code table} is {@code null}.
     */
    /* @NonNull */
    public static INetworkTable createNetworkTable(
        /* @NonNull */
        final ITable table )
    {
        return new NetworkTable( table );
    }
}
